add local validation / cleaner code / add background image according to trip
