+++ src/assets/stylesheets/reserve.css
@@ -7,7 +7,6 @@
 }
 
 .reserve-img {
-  background: url('https://rare-gallery.com/uploads/posts/558972-sky-high-resolution.jpg') no-repeat center center fixed;
   background-size: cover;
   height: 100vh;
   width: 100%;
